Reorganize blocks into a package
